Improve performance of parsing Packages/Sources files
Instead of reading the compressed version using gzreadline, read the
uncompressed version. In itself that does not improve performance,
but it allows to set the record separator to a blank line so we
can read an entire package section at a time instead of having to
build up sections manually. This results in a performance boost by
a factor of around 30.

git-svn-id: svn://svn.debian.org/debmirror/trunk@83 73d8de8a-19be-0310-9dbb-b31ba50de310