Make use of remote root directory more consistent: no more prefixes
The default leading "/" results in an redundant slash in urls.
Removing the "/" also makes it possible to simply double the ":" between
host name and root directory for rsync instead of requiring the user to
set --root with a leading ":".

Remove the leading "/" or ":" from parameters passed by users for
backwards compatibility.

git-svn-id: svn://svn.debian.org/debmirror/trunk@92 73d8de8a-19be-0310-9dbb-b31ba50de310