New option --rsync-extra to download additional files using rsync
Some files/directories on Debian mirrors are not listed in index files
and can even be said not to be part of the actual archive. For such
files rsync is the most natural transfer method as that avoids the need
of a complicated check which files have been added/changed/removed.

Supported with this commit are:
- mirror trace files (./project/trace/*)
- the directory ./doc plus ./README* files
- the directory ./indices
- the directory ./tools

Specified files/dircetories are automatically added to the "ignore" list
so that they don't get deleted during the cleanup stage.

git-svn-id: svn://svn.debian.org/debmirror/trunk@121 73d8de8a-19be-0310-9dbb-b31ba50de310