Don't delete Contents and Translation files with early mirror cleanup
With pre-mirror cleanup Contents and Translation files need to be
downloaded before the cleanup as otherwise they would be deleted
(and later re-downloaded) because they haven't been registered yet.

With post-mirror cleanup it's more neat to do all downloads together.

This could be simplified if we could register the files earlier.

git-svn-id: svn://svn.debian.org/debmirror/trunk@212 73d8de8a-19be-0310-9dbb-b31ba50de310