parse all types of checksums in the Release file
Changed check_file to named parameters so it can accept multiple
checksum types for a file, and choose which it prefers to check.

Changed %file_lists to use the same spelling for checksums that
the Release file uses, and the data in the hash for a given file
can now be passed directly into check_file.