Do not crash on i18n/Index with --debmarshal
The "i18n" subdirectory does not exist in the snapshot directory, so the call to
"link()" fails on the files below ".../i18n/".

In the past, this bug did not surface unless the "--i18n" option was also given.
With debmirror 2.12, "--i18n" is partly implied, so the crash happens
regardless.

With this patch, the target directory is guaranteed to exist before attepting to
link.

Also, "link_index_into_snapshot()" has been removed since it is not called
anywhere as of version 2.13.