debmirror: [PATCH] improvements to handling of gpgv subprocess
Instead of searching gpgv's output for a string "VALIDSIG", check its
exit status.  The section RETURN VALUE in gpgv(1) says:

       The  program  returns 0 if everything is fine, 1 if at least one signa‐
       ture was bad, and other error codes for fatal errors.

By contrast, there is nothing in gpgv's interface specification to
suggest that the string "VALIDSIG" is a public interface to be
relied upon.

Instead of re-running gpgv to display output to the user, capture the
standard output and standard error the first time the program is run,
and display those to the user if required.  Note that this may change
the ordering slightly: before, they could have been interleaved,
whereas now the standard output is displayed first, then the standard
error.

Instead of running the gpgv command via a string passed to a shell,
run it via an argument list passed to an exec-family syscall.  This
makes it possible to add additional user-supplied arguments to the
gpgv command without worrying about shell escaping.  In particular,
it would now be possible to add a --keyring option to debmirror,
which could be passed through to gpgv without worrying about escaping
any filenames from the shell.